+++ Tests/ShTests/MacTests.swift
@@ -28,6 +28,7 @@ final class MacTests: XCTestCase {
     XCTAssertNil(allOutput.terminationError)
     XCTAssertGreaterThan(allOutput.stdOut.count, 0)
     XCTAssertTrue(allOutput.stdOut.asTrimmedString()?.localizedCaseInsensitiveContains("com.apple.CoreSimulator") ?? false)
+    XCTAssertGreaterThan(allOutput.stdErr.count, 0)
     XCTAssertTrue(allOutput.stdErr.asTrimmedString()?.localizedCaseInsensitiveContains("Unknown binary with magic") ?? false)
   }
   
@@ -37,6 +38,7 @@ final class MacTests: XCTestCase {
     XCTAssertNil(allOutput.terminationError)
     XCTAssertGreaterThan(allOutput.stdOut.count, 0)
     XCTAssertTrue(allOutput.stdOut.asTrimmedString()?.localizedCaseInsensitiveContains("com.apple.CoreSimulator") ?? false)
+    XCTAssertGreaterThan(allOutput.stdErr.count, 0)
     XCTAssertTrue(allOutput.stdErr.asTrimmedString()?.localizedCaseInsensitiveContains("Unknown binary with magic") ?? false)
   }
